requirements.txt
aiogram==3.4.0
FastAPI==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
